Restructure rules.py into three modules
